cloud-functions/slide-bot/requirements.txt
google-api-python-client>=2.0.0
google-auth>=2.0.0
functions-framework>=3.0.0